+++ publiccms-by-gradle/src/com/publiccms/entities/log/LogTask.java
@@ -28,7 +28,7 @@ public class LogTask implements java.io.Serializable {
      */
     private static final long serialVersionUID = 1L;
     @MyColumn(title = "ID")
-    private Integer id;
+    private Long id;
     @MyColumn(title = "站点", condition = true)
     private int siteId;
     @MyColumn(title = "任务", condition = true)
@@ -52,7 +52,7 @@ public class LogTask implements java.io.Serializable {
         this.success = success;
     }
     
-    public LogTask(int id, int siteId, int taskId, Date begintime, Date endtime, boolean success) {
+    public LogTask(Long id, int siteId, int taskId, Date begintime, Date endtime, boolean success) {
         this.id = id;
         this.siteId = siteId;
         this.taskId = taskId;
@@ -73,11 +73,11 @@ public class LogTask implements java.io.Serializable {
     @Id
     @GeneratedValue(strategy = IDENTITY)
     @Column(name = "id", unique = true, nullable = false)
-    public Integer getId() {
+    public Long getId() {
         return this.id;
     }
 
-    public void setId(Integer id) {
+    public void setId(Long id) {
         this.id = id;
     }
