+++ publiccms-by-gradle/src/com/publiccms/logic/service/plugin/PluginLotteryUserService.java
@@ -18,7 +18,7 @@ import com.sanluan.common.handler.PageHandler;
 public class PluginLotteryUserService extends BaseService<PluginLotteryUser> {
 
     @Transactional(readOnly = true)
-    public PageHandler getPage(Integer lotteryId, Integer userId, Boolean winning, Date startCreateDate, Date endCreateDate,
+    public PageHandler getPage(Integer lotteryId, Long userId, Boolean winning, Date startCreateDate, Date endCreateDate,
             String orderType, Integer pageIndex, Integer pageSize) {
         return dao.getPage(lotteryId, userId, winning, startCreateDate, endCreateDate, orderType, pageIndex, pageSize);
     }
