+++ publiccms-by-gradle/src/com/publiccms/views/directive/cms/CmsContentListDirective.java
@@ -33,9 +33,9 @@ public class CmsContentListDirective extends AbstractTemplateDirective {
             }
         }
         PageHandler page = service.getPage(getSite(handler).getId(), status, handler.getInteger("categoryId"),
-                handler.getBoolean("containChild"), disabled, handler.getIntegerArray("modelId"), handler.getInteger("parentId"),
+                handler.getBoolean("containChild"), disabled, handler.getIntegerArray("modelId"), handler.getLong("parentId"),
                 emptyParent, handler.getBoolean("onlyUrl"), handler.getBoolean("hasImages"), handler.getBoolean("hasFiles"),
-                handler.getString("title"), handler.getInteger("userId"), handler.getInteger("checkUserId"),
+                handler.getString("title"), handler.getLong("userId"), handler.getLong("checkUserId"),
                 handler.getDate("startPublishDate"), endPublishDate, handler.getString("orderField"),
                 handler.getString("orderType"), handler.getInteger("pageIndex", 1), handler.getInteger("count", 30));
         handler.put("page", page).render();
