+++ publiccms-by-gradle/src/com/publiccms/views/directive/cms/CmsContentTagListDirective.java
@@ -17,8 +17,8 @@ public class CmsContentTagListDirective extends AbstractTemplateDirective {
 
     @Override
     public void execute(RenderHandler handler) throws IOException, Exception {
-        PageHandler page = service.getPage(handler.getInteger("tagId"), handler.getInteger("contentId"),
-                handler.getIntegerArray("contentIds"), handler.getInteger("pageIndex", 1),
+        PageHandler page = service.getPage(handler.getLong("tagId"), handler.getLong("contentId"),
+                handler.getLongArray("contentIds"), handler.getInteger("pageIndex", 1),
                 handler.getInteger("count", 30));
         handler.put("page", page).render();
     }
