+++ publiccms-by-gradle/src/com/publiccms/views/directive/cms/CmsWordDirective.java
@@ -20,11 +20,11 @@ public class CmsWordDirective extends AbstractTemplateDirective {
 
     @Override
     public void execute(RenderHandler handler) throws IOException, Exception {
-        Integer id = handler.getInteger("id");
+        Long id = handler.getLong("id");
         if (notEmpty(id)) {
             handler.put("object", service.getEntity(id)).render();
         } else {
-            Integer[] ids = handler.getIntegerArray("ids");
+            Long[] ids = handler.getLongArray("ids");
             if (notEmpty(ids)) {
                 List<CmsWord> entityList = service.getEntitys(ids);
                 Map<String, CmsWord> map = new LinkedHashMap<String, CmsWord>();
