+++ publiccms-by-gradle/src/com/publiccms/views/directive/sys/SysRoleUserListDirective.java
@@ -17,7 +17,7 @@ public class SysRoleUserListDirective extends AbstractTemplateDirective {
 
     @Override
     public void execute(RenderHandler handler) throws IOException, Exception {
-        PageHandler page = service.getPage(handler.getInteger("roleId"), handler.getInteger("userId"),
+        PageHandler page = service.getPage(handler.getInteger("roleId"), handler.getLong("userId"),
                 handler.getInteger("pageIndex", 1), handler.getInteger("count", 30));
         handler.put("page", page).render();
     }
